Update labels, change team resource to department resource
