+++ app-modules/ai/src/Filament/Resources/Prompts/Pages/CreatePrompt.php
@@ -103,7 +103,7 @@ class CreatePrompt extends CreateRecord
                             Select::make('confidential_prompt_teams')
                                 ->relationship('confidentialAccessTeams', 'name')
                                 ->preload()
-                                ->label('Teams')
+                                ->label('Departments')
                                 ->multiple()
                                 ->exists('teams', 'id')
                                 ->visible(fn (Get $get) => $get('is_confidential')),
