+++ app-modules/interaction/src/Filament/Resources/Interactions/Pages/CreateInteraction.php
@@ -118,7 +118,7 @@ class CreateInteraction extends CreateRecord
                             Select::make('interaction_confidential_teams')
                                 ->relationship('confidentialAccessTeams', 'name')
                                 ->preload()
-                                ->label('Teams')
+                                ->label('Departments')
                                 ->multiple()
                                 ->exists('teams', 'id')
                                 ->visible(fn (Get $get) => $get('is_confidential')),
