+++ app-modules/interaction/src/Filament/Resources/Interactions/Schemas/InteractionForm.php
@@ -126,7 +126,7 @@ class InteractionForm
                             Select::make('interaction_confidential_teams')
                                 ->relationship('confidentialAccessTeams', 'name')
                                 ->preload()
-                                ->label('Teams')
+                                ->label('Departments')
                                 ->multiple()
                                 ->exists('teams', 'id')
                                 ->visible(fn (Get $get) => $get('is_confidential')),
