+++ app-modules/meeting-center/src/Filament/Resources/BookingGroups/Pages/CreateBookingGroup.php
@@ -113,7 +113,7 @@ class CreateBookingGroup extends CreateRecord
                         ->preload()
                         ->live(),
                     Select::make('teams')
-                        ->label('Teams')
+                        ->label('Departments')
                         ->multiple()
                         ->relationship('teams', 'name')
                         ->searchable()
