+++ app-modules/project/src/Filament/Resources/Projects/Pages/ManageTasks.php
@@ -112,7 +112,7 @@ class ManageTasks extends ManageRelatedRecords
                                 Select::make('confidential_task_teams')
                                     ->relationship('confidentialAccessTeams', 'name')
                                     ->preload()
-                                    ->label('Teams')
+                                    ->label('Departments')
                                     ->multiple()
                                     ->exists('teams', 'id')
                                     ->visible(fn (Get $get) => $get('is_confidential')),
@@ -246,7 +246,7 @@ class ManageTasks extends ManageRelatedRecords
                                 Select::make('confidential_task_teams')
                                     ->relationship('confidentialAccessTeams', 'name')
                                     ->preload()
-                                    ->label('Teams')
+                                    ->label('Departments')
                                     ->multiple()
                                     ->exists('teams', 'id')
                                     ->visible(fn (Get $get) => $get('is_confidential')),
