+++ app-modules/report/src/Filament/Widgets/ProspectInteractionUsersTable.php
@@ -286,10 +286,10 @@ class ProspectInteractionUsersTable extends BaseWidget
                         }
                     }),
                 SelectFilter::make('team')
-                    ->label('Team')
+                    ->label('Department')
                     ->relationship('team', 'name')
                     ->multiple()
-                    ->placeholder('Select Team')
+                    ->placeholder('Select Department')
                     ->searchable()
                     ->preload(),
             ])
