+++ app-modules/student-data-model/src/Filament/Resources/Educatables/Pages/Concerns/CanManageEducatableTasks.php
@@ -104,7 +104,7 @@ trait CanManageEducatableTasks
                         Select::make('confidential_task_teams')
                             ->relationship('confidentialAccessTeams', 'name')
                             ->preload()
-                            ->label('Teams')
+                            ->label('Departments')
                             ->multiple()
                             ->exists('teams', 'id')
                             ->visible(fn (Get $get) => $get('is_confidential')),
